backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic[email]
pydantic-settings
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
aiofiles
reportlab